move `TaskContext` into `vm` module
